+++ changelog.md
@@ -1,5 +1,10 @@
 All notable changes to the Laravel Decomposer will be documented in this file
 
+## v1.2 (03-04-2017)
+- Now App & Other Package devs can also [add extra personal package or app specific stats](https://github.com/lubusIN/laravel-decomposer/wiki/Add-your-extra-stats) by using Decomposer as a dependency in their project
+- `getReportJson()` method added to get the same Decomposer report as JSON 
+- Fixed breaking of `getReportArray()` method
+
 ## v1.1.2 (19-03-2017)
 - Dependency improvements